auction_platform/api/serializers.py
from rest_framework import serializers
from auctions.models import Auction
from bids.models import Bid


class AuctionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Auction
        fields = '__all__'


class BidSerializer(serializers.ModelSerializer):
    class Meta:
        model = Bid
        fields = '__all__'
