requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2025.6
Django==6.0.2
django-crispy-forms==2.5
djangorestframework==3.16.1
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
